変更 : 重さによってスライムの速度が変わるようにしました

+++ Assets/MyAssets/InGame/Scripts/Slimes/SlimeParameters.cs
@@ -18,7 +18,7 @@ namespace Assets.MyAssets.InGame.Slimes
         /// スライムの重量
         /// ジャンプの高さなどに影響する
         /// </summary>
-        public int Weight;
+        public float Weight;
         
         /// <summary>
         /// スライムの大きさ
